Extract ShareCache persistence into the sharecache subpackage
